move protos to versioned mod
