+++ src/daft-proto/src/v1/mod.rs
@@ -1,2 +1,11 @@
 pub mod client;
 pub mod schema;
+
+#[allow(
+    clippy::must_use_candidate,
+    clippy::missing_errors_doc,
+    clippy::doc_markdown,
+    clippy::all,
+    clippy::pedantic
+)]
+pub mod protos;
