+++ src/daft-proto/src/v1/protos.rs
@@ -1,0 +1,11 @@
+//
+// !! PLEASE ENSURE THE MODULE STRUCTURE FOLLOWS PROTO DIR !!
+//
+
+pub mod echo {
+    tonic::include_proto!("echo");
+}
+
+pub mod schema {
+    tonic::include_proto!("schema");
+}
